Revert "change flags as well"
This reverts commit ed8897dd86b0dbfc80e6b8633f36d80c9bc91052.

+++ board/safety/safety_defaults.h
@@ -47,7 +47,7 @@ const safety_hooks nooutput_hooks = {
 // *** all output safety mode ***
 
 // Enables passthrough mode where relay is open and bus 0 gets forwarded to bus 2 and vice versa
-const uint32_t ALLOUTPUT_PARAM_PASSTHROUGH = 1;
+const uint16_t ALLOUTPUT_PARAM_PASSTHROUGH = 1;
 bool alloutput_passthrough = false;
 
 static const addr_checks* alloutput_init(uint32_t param) {
